Works for groups, figs in separate folders in ../reports/figs
Define groups to consider. In a small example, we will pick two groups, based on the treatment:

```{r}
useTreatment <- c("H","HD")
```

Output os in the /scripts/, same filename (05-DIABLO-master.html)  will be overwritten with each run. You should rename or move the html to preserve it.

Looking for a solution to automatize this.